rosen_backend/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
dnspython==2.2.1
Markdown==3.4.1
psycopg2==2.9.5
pymongo==4.3.3
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.5
